todo application

+++ delete.php
@@ -1,10 +1,14 @@
-<?php 
-  include('./db_config.php');
-  $id = $_GET['deleteid'];
-  $sql = "DELETE FROM todo_list WHERE id= $id";
-  $result = $conn->query($sql);
-  if($result){
-    header('Location: ./index.php');
-  }
+<?php
+    include('./db_config.php');
+    if(isset($_GET['deleteid'])){
+        $id = $_GET['deleteid'];
 
+        $sql = "DELETE FROM todo_list WHERE id = $id";
+        $result = mysqli_query($conn,$sql);
+        if($result){
+          header('Location: index.php');
+        }else{
+          die($conn -> connect_error);
+        }
+    }
 ?>
